requirements.txt
fitz==0.0.1.dev2
Pillow==10.4.0
pytesseract==0.3.13
python_docx==1.1.2
reportlab==4.2.2
